+++ source/main/cardano/CardanoWalletLauncher.js
@@ -38,6 +38,7 @@ export type WalletOpts = {
   walletLogFile: WriteStream,
   cliBin: string,
   isStaging: boolean,
+  smashUrl?: string,
 };
 
 export async function CardanoWalletLauncher(walletOpts: WalletOpts): Launcher {
@@ -56,6 +57,7 @@ export async function CardanoWalletLauncher(walletOpts: WalletOpts): Launcher {
     walletLogFile,
     cliBin,
     isStaging,
+    smashUrl,
   } = walletOpts;
   // TODO: Update launcher config to pass number
   const syncToleranceSeconds = parseInt(syncTolerance.replace('s', ''), 10);
@@ -127,6 +129,10 @@ export async function CardanoWalletLauncher(walletOpts: WalletOpts): Launcher {
         // All clusters not flagged as staging except for Mainnet are treated as "Testnets"
         launcherConfig.networkName = TESTNET;
         logger.info('Launching Wallet with --testnet flag');
+      }
+      if (smashUrl) {
+        logger.info('Launching Wallet with --smash-url flag', { smashUrl });
+        merge(launcherConfig, { smashUrl });
       }
       merge(launcherConfig, { nodeConfig, tlsConfiguration });
       break;